Removed error-handling
Because arrays are passed as pointers in C, I cannot check the size of the array without erroring in C. Might need to check before passing, or add some error handling later.

Also changed syntax error with arrays being passed into a function.

Removed initRC call, as it only needs to be called ONCE in the code, otherwise the PWM outputs will be deleted upon each call. I need to add initRC to main. 

Fixed PEMDAS order-error with scaling factor (had division occur when it should not have).

And changed DelayMs to msDelay (which is the proper function call).